правки в ЛК

+++ frontend/web/vue/src/components/AllCompany.vue
@@ -31,6 +31,7 @@
                                 <v-btn outline small fab
                                        class="edit-btn"
                                        type="button"
+                                       title="Редактировать"
                                 >
                                     <v-icon>edit</v-icon>
 
@@ -49,6 +50,7 @@
                             <v-btn outline small fab
                                    class="edit-btn"
                                    type="button"
+                                   title="Удалить"
                                    @click="removeResume(index, item.id)"
                             >
                                 <v-icon>delete</v-icon>
@@ -123,21 +125,33 @@
         },
         methods: {
             removeResume(index, resumeId) {
-                this.getAllCompany.splice(index, 1);
-                this.$http.delete(`${process.env.VUE_APP_API_URL}/request/company/` + resumeId)
-                    .then(response => {
-                        return response;
-                        }, response => {
-                        this.$swal({
-                            toast: true,
-                            position: 'bottom-end',
-                            showConfirmButton: false,
-                            timer: 4000,
-                            type: 'error',
-                            title: response.data.message
-                        })
-                        }
-                    );
+                this.$swal({
+                    title: 'Вы точно хотите удалить компанию?',
+                    type: 'warning',
+                    showCancelButton: true,
+                    confirmButtonColor: '#3085d6',
+                    cancelButtonColor: '#d33',
+                    confirmButtonText: 'Да',
+                    cancelButtonText: 'Нет'
+                }).then((result) => {
+                    if (result.value) {
+                        this.getAllCompany.splice(index, 1);
+                        this.$http.delete(`${process.env.VUE_APP_API_URL}/request/company/` + resumeId)
+                            .then(response => {
+                                    return response;
+                                }, response => {
+                                    this.$swal({
+                                        toast: true,
+                                        position: 'bottom-end',
+                                        showConfirmButton: false,
+                                        timer: 4000,
+                                        type: 'error',
+                                        title: response.data.message
+                                    })
+                                }
+                            );
+                    }
+                });
             },
             changePage(paginationCurrentPage) {
                 this.$http.get(`${process.env.VUE_APP_API_URL}/request/company/my-index?page=` + paginationCurrentPage)
